# Conflicts:
#	cmd/state/exec3/historical_trace_worker.go
#	cmd/state/exec3/state.go
#	cmd/state/exec3/trace_worker.go
#	core/blockchain.go
#	core/state_processor.go
#	core/state_transition.go
#	eth/backend.go
#	eth/stagedsync/exec3.go
#	eth/tracers/internal/tracetest/calltrace_test.go
#	eth/tracers/internal/tracetest/prestate_test.go
#	eth/tracers/native/call.go
#	eth/tracers/tracers_test.go
#	turbo/jsonrpc/otterscan_api.go
#	turbo/jsonrpc/otterscan_search_trace.go
#	turbo/jsonrpc/otterscan_trace_transaction.go
#	turbo/jsonrpc/trace_adhoc.go
#	turbo/jsonrpc/trace_filtering.go
#	turbo/jsonrpc/tracing.go
#	turbo/stages/stageloop.go
#	turbo/transactions/tracing.go

+++ turbo/execution/eth1/ethereum_execution.go
@@ -167,12 +167,13 @@ func (e *EthereumExecutionModule) unwindToCommonCanonical(tx kv.RwTx, header *ty
 	currentHeader := header
 
 	for isCanonical, err := e.isCanonicalHash(e.bacgroundCtx, tx, currentHeader.Hash()); !isCanonical && err == nil; isCanonical, err = e.isCanonicalHash(e.bacgroundCtx, tx, currentHeader.Hash()) {
-		currentHeader, err = e.getHeader(e.bacgroundCtx, tx, currentHeader.ParentHash, currentHeader.Number.Uint64()-1)
+		parentBlockHash, parentBlockNum := currentHeader.ParentHash, currentHeader.Number.Uint64()-1
+		currentHeader, err = e.getHeader(e.bacgroundCtx, tx, parentBlockHash, parentBlockNum)
 		if err != nil {
 			return err
 		}
 		if currentHeader == nil {
-			return fmt.Errorf("header %v not found", currentHeader.Hash())
+			return fmt.Errorf("header %d, %x not found", parentBlockNum, parentBlockHash)
 		}
 	}
 	if err := e.hook.BeforeRun(tx, true); err != nil {
@@ -291,6 +292,8 @@ func (e *EthereumExecutionModule) purgeBadChain(ctx context.Context, tx kv.RwTx,
 		return err
 	}
 
+	dbHeadHash := rawdb.ReadHeadBlockHash(tx)
+
 	currentHash := headHash
 	currentNumber := *tip
 	for currentHash != latestValidHash {
@@ -298,6 +301,13 @@ func (e *EthereumExecutionModule) purgeBadChain(ctx context.Context, tx kv.RwTx,
 		if err != nil {
 			return err
 		}
+
+		// TODO: find a better way to handle this
+		if currentHash == dbHeadHash {
+			// We can't delete the head block stored in the database as that is our canonical reconnection point.
+			return nil
+		}
+
 		rawdb.DeleteHeader(tx, currentHash, currentNumber)
 		currentHash = currentHeader.ParentHash
 		currentNumber--
